Remove misconfigured guard
In the example for running the linter on all files, the 

```
    if: steps.filter.outputs.eslintrc == 'false'
```

doesn't make sense to me. When I tried the example on my project, the guard prevented the eslint action from running on any PR where a `.eslint*` file had changed. The  `teps.filter.outputs.eslintrc` condition is the same condition as in the `all-files` paremeter, so it means the step would never run in the case where `all-files` would have been true. 

Is the guard maybe left over from an earlier example where two different steps were run, depending on whether eslint config had changed?